move related files into the new search package and make necessary changes
