+++ src/main/java/hello/controllers/EarthquakesController.java
@@ -16,7 +16,7 @@ import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
 import hello.geojson.FeatureCollection;
 
 import hello.services.EarthquakeQueryService;
-import hello.EqSearch;
+import hello.searches.EqSearch;
 
 @Controller
 public class EarthquakesController {
